split pr-label.mts file
