+++ nodescripts/pr-label/constants.mts
@@ -1,0 +1,27 @@
+export const REVIEWER_USERNAMES = new Set([
+  'frankfe-amsterdam',
+  'RikSchefferAmsterdam',
+  'WouterAms',
+  'fhaver-amsterdam',
+  'jjbeekman',
+])
+export const COPILOT_LOGINS = new Set(['github-copilot[bot]'])
+export const REVIEWED_STATES = new Set([
+  'APPROVED',
+  'CHANGES_REQUESTED',
+  'DISMISSED',
+])
+export const MODULE_LABEL_PREFIX = 'module:'
+export const MODULE_LABEL_COLOR = '0366d6'
+export const GENERAL_LABEL_COLOR = '036d66'
+export const COPILOT_READY_LABEL = 'Copilot ready'
+export const COPILOT_READY_LABEL_COLOR = '05b103'
+export const REVIEWED_LABEL = 'Code reviewed'
+export const REVIEWED_LABEL_COLOR = '05b103'
+export const HTTP_STATUS_UNPROCESSABLE_ENTITY = 422
+export const MODULE_PATH_RE = /^src\/modules\/([^/]+)\//
+export const CHANGES_HEADING = '# Changes'
+export const TEST_INSTRUCTIONS_HEADING = '## Test instructions'
+export const REVIEWED_CHANGES_HEADING = '### Reviewed changes'
+export const DETAILS_OPEN_TAG = '<details>'
+export const DETAILS_CLOSE_TAG = '</details>'
